+++ examples/design-method-simulate-DualDesign.R
@@ -11,7 +11,7 @@ model <- DualEndpointRW(mean = c(0, 1),
                         use_log_dose=TRUE,
                         ref_dose=2,
                         rho = c(a=1, b=1),
-                        smooth="RW1")
+                        rw1 = TRUE)
 
 # Choose the rule for selecting the next dose
 myNextBest <- NextBestDualEndpoint(target=c(0.9, 1),