+++ NEWS.md
@@ -1,4 +1,5 @@
 # Version 1.0.9000.9000
+* Replaced `smooth` character vector argument of `DualEndpointRW()` to `rw1` flag. 
 * Added new `ModelParamsNormal` class to represent parameters of bivariate normal
   distribution.
 * Added new `doseFunction` and `probFunction` functions which return dose